+++ Exodus-Test-Server/src/utopia/exodus/test/rest/servlet/ApiServlet.scala
@@ -5,11 +5,7 @@ import javax.servlet.http.{HttpServlet, HttpServletRequest, HttpServletResponse}
 import utopia.access.http.Method
 import utopia.access.http.Status.BadRequest
 import utopia.bunnymunch.jawn.JsonBunny
-import utopia.exodus.rest.resource.description.{DescriptionRolesNode, LanguageFamiliaritiesNode, LanguagesNode, RolesNode, TasksNode}
-import utopia.exodus.rest.resource.device.DevicesNode
-import utopia.exodus.rest.resource.email.EmailsNode
-import utopia.exodus.rest.resource.organization.OrganizationsNode
-import utopia.exodus.rest.resource.user.{QuestSessionsNode, UsersNode}
+import utopia.exodus.rest.resource.ExodusResources
 import utopia.exodus.rest.util.AuthorizedContext
 import utopia.exodus.util.ExodusContext
 import utopia.flow.async.ThreadPool
@@ -48,6 +44,7 @@ class ApiServlet extends HttpServlet
 	
 	val dbSettings = JsonBunny.munchPath("settings/exodus-db-settings.json").map { _.getModel }
 	
+	// FIXME: Look up default user scopes (second argument list)
 	ExodusContext.setup(exc, connectionPool,
 		dbSettings.toOption.flatMap { _("db_name").string }.getOrElse("exodus_db")) { (error, message) =>
 		println(message)
@@ -71,11 +68,7 @@ class ApiServlet extends HttpServlet
 	private implicit val serverSettings: ServerSettings = ServerSettings("http://localhost:9999")
 	private implicit val jsonParser: JsonParser = JsonBunny
 	
-	private val handler = new RequestHandler(Map("v1" -> Vector(
-			DescriptionRolesNode.public, LanguagesNode.public, LanguageFamiliaritiesNode.public, RolesNode, TasksNode,
-			UsersNode.forApiKey, DevicesNode, OrganizationsNode, QuestSessionsNode,
-			EmailsNode.forApiKey
-		)),
+	private val handler = new RequestHandler(Map("v1" -> ExodusResources.all),
 		Some(Path("exodus", "api")), r => AuthorizedContext(r))